add *repository for entities

+++ src/main/java/io/everyonecodes/project_module/entity/Dungeon.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 @Data
 public class Dungeon {
-    private final Integer id;
+    private final Long id;
     private String name;
     private String description;
     private Integer difficultyLevel;
-    private Integer monsterId;
-    private List<Integer> heroIds;
+    private Long monsterId;
+    private List<Long> heroIds;
 }
